+++ src/main/java/roomescape/controller/ThemeController.java
@@ -11,20 +11,26 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.Theme;
 import roomescape.dto.PopularThemeResponse;
 import roomescape.dto.ReservationTimeStatusResponse;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
+import roomescape.service.ReservationTimeService;
 import roomescape.service.ThemeService;
 
 @RestController
 public class ThemeController {
 
     private final ThemeService themeService;
+    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationTimeService reservationTimeService;
 
-    public ThemeController(ThemeService themeService) {
+    public ThemeController(ThemeService themeService, ReservationTimeDao reservationTimeDao, ReservationTimeService reservationTimeService) {
         this.themeService = themeService;
+        this.reservationTimeDao = reservationTimeDao;
+        this.reservationTimeService = reservationTimeService;
     }
 
     @GetMapping("/themes")
@@ -50,7 +56,7 @@ public class ThemeController {
 
     @GetMapping("/themes/{themeId}/reservation-times")
     public List<ReservationTimeStatusResponse> read(@RequestParam LocalDate date, @PathVariable Long themeId) {
-        return themeService.findReservationTimeByDateAndThemeId(date, themeId);
+        return reservationTimeService.findReservationTimeByDateAndThemeId(date, themeId);
     }
 
     @GetMapping("/themes/popular")
