+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,11 +1,14 @@
 package roomescape.service;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeStatusResponse;
 
 @Service
 public class ReservationTimeService {
@@ -32,6 +35,20 @@ public class ReservationTimeService {
     public void deleteById(Long id) {
         validateHasTime(id);
         reservationTimeDao.deleteById(id);
+    }
+
+    public List<ReservationTimeStatusResponse> findReservationTimeByDateAndThemeId(LocalDate date, Long themeId) {
+        List<ReservationTime> reservationTimes = reservationTimeDao.findAll();
+        List<Long> timeIds = reservationDao.findReservedTimeIdsByDateAndThemeId(date, themeId);
+
+        LocalDateTime now = LocalDateTime.now();
+        return reservationTimes.stream()
+                .map(reservationTime -> {
+                    boolean available = !timeIds.contains(reservationTime.getId())
+                            && !reservationTime.isPast(date, now);
+                    return ReservationTimeStatusResponse.of(reservationTime, available);
+                })
+                .toList();
     }
 
     private void validateHasTime(Long id) {
